requirements.txt
flask
python-dotenv
azure-storage-blob
azure-storage-file-share
gunicorn
